+++ modules/apps/document-library/document-library-repository-portlet-file-repository-impl/build.gradle
@@ -1,2 +1,8 @@
 dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component", version: "1.3.0"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	compileOnly project(":core:petra:petra-function")
+	compileOnly project(":core:petra:petra-string")
 }
